+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `deltoids`: fix bug where diff did not show when adding a line at end of file
+- `deltoids`: fix bug where newly added files showed nested ancestor scope boxes
 
 ## [0.1.0] - 2025-04-23